[Bug 6234]:  Fixing flags
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@75170 9d14d920-75d5-0310-8bd7-a52572981aee

+++ tools/trc/lib/trc_update.h
@@ -60,7 +60,7 @@ typedef TAILQ_HEAD(trc_update_tags_logs,
 
 /** Context of TRC Update tool */
 typedef struct trc_update_ctx {
-    unsigned int             flags;           /**< Flags */
+    uint64_t                 flags;           /**< Flags */
     te_trc_db               *db;              /**< TRC DB */
     unsigned int             db_uid;          /**< TRC DB user ID */
     tqh_strings              test_names;      /**< Test paths */
